Improve history widget layout and fix overflow in finish_line
- Change App::num field type from u8 to u64
- Fix finish_line to guard against overflow when len >= width
- Refactor history widget: update field types and dimensions for date, time, label and value
- Add branching by HistoryType (Hour, Day, Month) in next_item
- Update date/time string formatting in update(), comment out seconds
- Right-align value display with flow or 'None'

+++ src/apps.rs
@@ -38,7 +38,7 @@ pub struct HistoryState {
 
 pub struct App {
     pub text: &'static str,
-    pub num: u8,
+    pub num: u64,
     pub label_title: &'static str,
     pub label_value: String,
     pub datetime: PrimitiveDateTime,
